tests/requirements.txt
# ITSM自动化测试依赖
# ITSM Automated Test Dependencies

# 测试框架
pytest>=7.4.0
pytest-html>=4.1.0
pytest-xdist>=3.5.0
pytest-timeout>=2.2.0
pytest-cov>=4.1.0
pytest-playwright>=0.4.0
pytest-repeat>=0.9.0

# API测试
requests>=2.31.0
urllib3>=2.0.0

# 数据库测试
psycopg2-binary>=2.9.9

# UI测试
playwright>=1.40.0

# 工具库
python-dotenv>=1.0.0
pyyaml>=6.0.1
colorama>=0.4.6
tabulate>=0.9.0

# 报告生成
jinja2>=3.1.2

# 异步测试
pytest-asyncio>=0.21.0
